% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_single_par.R
\name{calc_change_all_pars_sH1N1_H7N9_CI}
\alias{calc_change_all_pars_sH1N1_H7N9_CI}
\title{calculate how the cellular infection parameters change when model parameters
are changed from the value for sH1N1-WT to that for H7N9-WT}
\usage{
calc_change_all_pars_sH1N1_H7N9_CI()
}
\value{
a data frame.
sum_stat is the name of the cellular infection parameter: "log10_R_0", "gen_time" or "r".
par_name is the name of the model parameter that is changed: 
"p_inf", "delta", "c_inf", "k1", or "beta_inf"
2.5%, 50% and 97.5% are the quantiles of the percentage change in the cellular
infection parameter when the value of the model parameter is changed from
that of sH1N1-WT to that of H7N9-WT.
}
\description{
calculate how the cellular infection parameters change when model parameters
are changed from the value for sH1N1-WT to that for H7N9-WT
}
