% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_generation_distribution.R
\name{calc_gen_dist_vary_params}
\alias{calc_gen_dist_vary_params}
\title{calculate the generation number distribution as cellular infection parameters are varied}
\usage{
calc_gen_dist_vary_params(constant_param)
}
\arguments{
\item{constant_param}{character vector: "log10_R_0", "gen_time" or "r"}
}
\value{
a list with three objects.
plot_df is a data frame with three columns. gen is the generation number (1:20),
V is the proportion of infectious virions belonging to that generation at 
the time of peak viral load, and infection_par_value is the value of the
independent variable (which depends on constant_param).
all_pars is a list with three elements.
independent is a vector of values for the independent parameter,
dependent is a vector of values for the dependent parameter,
and constant is a vector of (repeating) values for the constant parameter.
Which parameter out of log10_R_0, gen_time and r is independent, dependent
or constant depends on constant_param.
sf is the number of significant figures to which the values in all_pars are
given (used for plotting later)
}
\description{
calculate the generation number distribution as cellular infection parameters are varied
}
