% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_generation_distribution.R
\name{calc_generations_at_time_fixed}
\alias{calc_generations_at_time_fixed}
\title{calculation for Fig. S5
for either the model where the viral load exponentially increases then stays constant,
or the model where the viral load exponentially increases then exponentially
decreases at the same rate, calculate the number of virions in each generation at a given time}
\usage{
calc_generations_at_time_fixed(time, gen_time = 1, V_0, R_0, I_max,
  viral_load_fn)
}
\arguments{
\item{time}{numeric vector of length 1.  time at which to calculate the number of virions in each generation}

\item{gen_time}{numeric vector of length 1.  fixed generation time}

\item{V_0}{numeric vector of length 1.  Inoculum size}

\item{R_0}{numeric vector of length 1.  Basic reproduction number}

\item{I_max}{numeric vector of length 1.  Maximum number of infected cells}

\item{viral_load_function}{function.
only calc_generations_fixed_growth_until_I_max is implemented currently}
}
\value{
numeric vector containing the number of virions in each generation at a given time.
}
\description{
calculation for Fig. S5
for either the model where the viral load exponentially increases then stays constant,
or the model where the viral load exponentially increases then exponentially
decreases at the same rate, calculate the number of virions in each generation at a given time
}
