% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_generation_distribution.R
\name{calc_generations_closure}
\alias{calc_generations_closure}
\title{make a function to calculate the generation number distribution at the time 
of peak infectious viral load, for a given set of parameters}
\usage{
calc_generations_closure(wash)
}
\arguments{
\item{wash}{logical.  If FALSE, exclude wash step for multi-cycle experiment.
If TRUE, do whatever the original multi-cycle experiment did.}
}
\value{
a function
}
\description{
make a function to calculate the generation number distribution at the time 
of peak infectious viral load, for a given set of parameters
}
