% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_generation_distribution.R
\name{calc_generations_fixed_growth_until_I_max}
\alias{calc_generations_fixed_growth_until_I_max}
\title{for a model where the viral load exponentially increases then stays constant, 
calculate the number of virions in each generation}
\usage{
calc_generations_fixed_growth_until_I_max(n_gens, V_0, R_0, I_max)
}
\arguments{
\item{n_gens}{numeric vector of length 1. number of generations to calculate up until}

\item{V_0}{numeric vector of length 1.  Inoculum size}

\item{R_0}{numeric vector of length 1.  Basic reproduction number}

\item{I_max}{numeric vector of length 1.  Maximum number of infected cells}
}
\value{
numeric vector of length n_gens.  Number of virions in each generation
}
\description{
for a model where the viral load exponentially increases then stays constant, 
calculate the number of virions in each generation
}
