% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_generation_distribution.R
\name{calc_mean_gen_no}
\alias{calc_mean_gen_no}
\title{calculate the mean generation number at the time of peak infectious viral load,
for a given set of parameters}
\usage{
calc_mean_gen_no(pars, wash)
}
\arguments{
\item{pars}{a named vector of parmaeter values}

\item{wash}{logical.  If FALSE, exclude wash step for multi-cycle experiment.
If TRUE, do whatever the original multi-cycle experiment did.}
}
\value{
a numeric vector of length 1: the mean generation number for that parameter set
}
\description{
calculate the mean generation number at the time of peak infectious viral load,
for a given set of parameters
}
