% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_generation_distribution.R
\name{calc_mean_gen_no_by_strain}
\alias{calc_mean_gen_no_by_strain}
\title{calculate the distribution of the mean generation number over the posterior
for a given strain}
\usage{
calc_mean_gen_no_by_strain(strain, sensitivity)
}
\arguments{
\item{strain}{character: sH1N1, pH1N1, H5N1, H7N9 for the WT strains;
ENG195, Turkey05 for pH1N1-PR8 and H5N1-PR8 respectively}

\item{sensitivity}{character: choice of "main" (model in main paper),
"n_L_n_I" (changing the number of latent and infectious stages), 
"loss" (ignoring loss of virus due to entry into target cells),
"mc" (use multi-cycle data only),
"linear_p" (production rate of virions increases linearly with time)}
}
\value{
a vector of samples from the distribution of the mean generation number
}
\description{
calculate the distribution of the mean generation number over the posterior
for a given strain
}
