% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_calc_parameters.R
\name{calc_mean_gen_time}
\alias{calc_mean_gen_time}
\title{calculate the mean generation time given model parameters}
\usage{
calc_mean_gen_time(c1, tau_E, tau_I, n_L, n_I, p_over_time = "constant")
}
\arguments{
\item{c1}{rate of loss of free virus (due to loss of infectivity + entering target cell)}

\item{tau_E}{mean latent period}

\item{tau_I}{mean infectious period}

\item{n_L}{number of latent stages}

\item{n_I}{number of infectious stages}

\item{p_over_time}{"constant" for constant rate of virus production over
infected cell lifetime, "linear" for linearly increasing rate of virus production
over infected cell lifetime}
}
\value{
numeric vector of length 1. mean generation time
}
\description{
calculate the mean generation time given model parameters
}
