% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_calc_parameters.R
\name{calc_mean_gen_time_wrapper}
\alias{calc_mean_gen_time_wrapper}
\title{calculate the mean generation time given named vector of parameter values}
\usage{
calc_mean_gen_time_wrapper(pars)
}
\arguments{
\item{pars}{named vector of parameter values}
}
\value{
mean generation time
}
\description{
calculate the mean generation time given named vector of parameter values
}
