% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_calc_parameters.R
\name{calc_R_0}
\alias{calc_R_0}
\title{calculates R_0 for the TEIV model with Erlang-distributed latent and infectious periods}
\usage{
calc_R_0(pars, p_over_time = "constant")
}
\arguments{
\item{pars}{named vector of parameter values}

\item{p_over_time}{"constant" for constant rate of virus production over
infected cell lifetime, "linear" for linearly increasing rate of virus production
over infected cell lifetime}
}
\value{
R_0: numeric vector of length 1
}
\description{
calculates R_0 for the TEIV model with Erlang-distributed latent and infectious periods
}
