% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_calc_parameters.R
\name{calc_r_teiv_stages}
\alias{calc_r_teiv_stages}
\title{calculates r for the TEIV model with Erlang-distributed latent and infectious periods}
\usage{
calc_r_teiv_stages(pars, p_over_time = "constant", warn_if_neg = FALSE)
}
\arguments{
\item{pars}{named vector of parameter values}

\item{p_over_time}{"constant" for constant rate of virus production over
infected cell lifetime, "linear" for linearly increasing rate of virus production
over infected cell lifetime}

\item{warn_if_neg}{logical. if TRUE, throw warning if a negative value of r results}
}
\value{
r: numeric vector of length 1
}
\description{
calculates r for the TEIV model with Erlang-distributed latent and infectious periods
}
