% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_for_inference.R
\name{CREATE_POSTERIOR_FUNC}
\alias{CREATE_POSTERIOR_FUNC}
\title{creates a closure to calculate the likelihood and model predictions for given
parameter values}
\usage{
CREATE_POSTERIOR_FUNC(parTab, data, PRIOR_FUNC, p_over_time = "constant")
}
\arguments{
\item{parTab}{data frame of model parameters created by specify_parameters}

\item{data}{data frame of data used for fitting in MCMC}

\item{PRIOR_FUNC}{NULL for uniform prior, otherwise closure specifying prior}

\item{p_over_time}{"constant" for constant rate of virus production over
infected cell lifetime, "linear" for linearly increasing rate of virus production
over infected cell lifetime}
}
\value{
closure to calculate likelihood and model predictions for given
parameter values
}
\description{
creates a closure to calculate the likelihood and model predictions for given
parameter values
}
