% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{define_palette}
\alias{define_palette}
\title{define colour palette for plotting}
\usage{
define_palette(data_source)
}
\arguments{
\item{data_source}{character: WT or PR8}
}
\value{
character vector of colour hexes
}
\description{
define colour palette for plotting
}
