% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_filenames.R
\name{determine_data_source_for_strain}
\alias{determine_data_source_for_strain}
\title{extract name of data source from strain name}
\usage{
determine_data_source_for_strain(strain)
}
\arguments{
\item{strain}{name of strain}
}
\value{
character vector of length 1
}
\description{
extract name of data source from strain name
}
