% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_filenames.R
\name{determine_strains_for_exp}
\alias{determine_strains_for_exp}
\title{list all strains for a data source}
\usage{
determine_strains_for_exp(data_source, w_data_source = FALSE)
}
\arguments{
\item{data_source}{character string: "WT", "PR8" or "all"}

\item{w_data_source}{logical. if TRUE, append data source to strain name
(compulsory if data_source == "all")}
}
\value{
character vector containing strain names
}
\description{
list all strains for a data source
}
