% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_summary_stats.R
\name{extract_cell_infect_params}
\alias{extract_cell_infect_params}
\title{calculate and save cellular infection parameters for MCMC chain}
\usage{
extract_cell_infect_params(data_dir)
}
\arguments{
\item{data_dir}{directory where 1.RData lives}
}
\value{
3xn matrix where each row corresponds to a cellular infection parameter and n is
the number of MCMC iterations
}
\description{
calculate and save cellular infection parameters for MCMC chain
}
