% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_for_inference.R
\name{gen_summary_statistics_fn}
\alias{gen_summary_statistics_fn}
\title{create closure to calculate summary statistics and provide bounds for them}
\usage{
gen_summary_statistics_fn(parTab, startTab, p_over_time)
}
\arguments{
\item{parTab}{data frame of model parameters created by specify_parameters}

\item{startTab}{data frame of smae format as parTab, storing starting
parameter values}

\item{p_over_time}{"constant" for constant rate of virus production over
infected cell lifetime, "linear" for linearly increasing rate of virus production
over infected cell lifetime}
}
\value{
a list with elements
"calc_summary": closure to calculate
}
\description{
create closure to calculate summary statistics and provide bounds for them
}
