% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_helpers.R
\name{get_MCMC_chain}
\alias{get_MCMC_chain}
\title{extract chain from MCMC output}
\usage{
get_MCMC_chain(filename_in)
}
\arguments{
\item{filename_in}{character vector of length 1: name of .RData file outputted by MCMC}
}
\value{
chain data table containing MCMC runs
}
\description{
\code{get_MCMC_chain} extracts chain from MCMC output, discarding burn-in
}
