% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_filenames.R
\name{get_p_over_time_from_sensitivity}
\alias{get_p_over_time_from_sensitivity}
\title{extract the value of p_over_time from the sensitivity input parameter}
\usage{
get_p_over_time_from_sensitivity(sensitivity)
}
\arguments{
\item{sensitivity}{character: choice of "main" (model in main paper),
"n_L_n_I" (changing the number of latent and infectious stages), 
"loss" (ignoring loss of virus due to entry into target cells),
"mc" (use multi-cycle data only),
"linear_p" (production rate of virions increases linearly with time)}
}
\value{
"linear" or "constant"
}
\description{
extract the value of p_over_time from the sensitivity input parameter
}
