% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_for_inference.R
\name{get_prediction_ci_df_all}
\alias{get_prediction_ci_df_all}
\title{extract model predictions form MCMC chain and calculate CIs}
\usage{
get_prediction_ci_df_all(chain, data_list)
}
\arguments{
\item{chain}{MCMC chain}

\item{data_list}{data used to fit model}
}
\value{
list of same lenght as data_list.  Each element contains:
data_df: the corresponding element of data_list
V_ci: the predicted viral load corresponding to that data
}
\description{
extract model predictions form MCMC chain and calculate CIs
}
