% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_filenames.R
\name{get_strain_dir}
\alias{get_strain_dir}
\title{get directory names containing MCMC output for strains}
\usage{
get_strain_dir(dirs, strains)
}
\arguments{
\item{dirs}{output of set_dirs}

\item{strains}{character vector of strain names}
}
\value{
character vector of same length as strains
}
\description{
get directory names containing MCMC output for strains
}
