% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_for_inference.R
\name{has_data_outer}
\alias{has_data_outer}
\title{checks if any string in the vector inc_data contains
any of the strings in the vector str}
\usage{
has_data_outer(str, inc_data)
}
\arguments{
\item{str}{character vector}

\item{inc_data}{character vector}
}
\value{
logical vector of length 1
}
\description{
checks if any string in the vector inc_data contains
any of the strings in the vector str
}
