% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make_corr_table}
\alias{make_corr_table}
\title{make correlation table for supplementary csv}
\usage{
make_corr_table(strain)
}
\arguments{
\item{strain}{character: sH1N1, pH1N1, H5N1, H7N9 for the WT strains;
ENG195, Turkey05 for pH1N1-PR8 and H5N1-PR8 respectively}
}
\value{
a matrix
}
\description{
make correlation table for supplementary csv
}
