% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make_Fig3}
\alias{make_Fig3}
\title{make Fig 3A-B}
\usage{
make_Fig3(sensitivity)
}
\arguments{
\item{sensitivity}{character: choice of "main" (model in main paper),
"n_L_n_I" (changing the number of latent and infectious stages), 
"loss" (ignoring loss of virus due to entry into target cells),
"mc" (use multi-cycle data only),
"linear_p" (production rate of virions increases linearly with time)}
}
\value{
a list of ggplot objects
}
\description{
make Fig 3A-B
}
