% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_for_inference.R
\name{make_inc_data}
\alias{make_inc_data}
\title{make character vector of data to infer/predict given inc_data_logical vector}
\usage{
make_inc_data(inc_data_logical = rep(T, 5), predict = FALSE)
}
\arguments{
\item{inc_data_logical}{logical vector of length 5.  indicates which of
\code{c("sc_inf", "sc_tot", "mc_inf", "mc_tot", "mock")} we have data for}

\item{predict}{logical.  If TRUE, if say data for "sc_inf" is included,
"sc_tot" will be included in the output also.  If FALSE, directly
translate inc_data_logical into character vector}
}
\value{
character vector
}
\description{
make character vector of data to infer/predict given inc_data_logical vector
}
