% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_summary_stats.R
\name{pairwise_stats_tests}
\alias{pairwise_stats_tests}
\title{test whether the ratio of a summary statistic between two strains
differs from 1}
\usage{
pairwise_stats_tests(data_dir, sum_stat, data_source)
}
\arguments{
\item{data_dir}{the n directories where the 1.RData files live}

\item{sum_stat}{character vector of length 1.  if "cell_infect_params", plot generation
time, R_0 and initial growth rate.
if "mean_gen_no", plot the mean generation number.}

\item{data_source}{character: WT or PR8}
}
\value{
a list of length n, where n is the number of pair combinations of data_dir.
each element contains a list with the elements prctile_ratios and p_value.
prctile_ratio is a matrix where each column corresponds to a summary statistic,
and the rows give the 2.5, 50 and 97.5th percentiles of the ratio of each
statistic between the two strains.
p_value is a named vector where each element is twice the proportion of the
sampled ratios greater than 1, or the proportion of the sampled ratios smaller
than 1 (whichever is the smaller)
}
\description{
test whether the ratio of a summary statistic between two strains
differs from 1
}
