% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_bivariate_density_multiple_strains}
\alias{plot_bivariate_density_multiple_strains}
\title{plot the bivariate density plot for R_0 and gen_time for multiple strains on the same plot}
\usage{
plot_bivariate_density_multiple_strains(data_source, sensitivity)
}
\arguments{
\item{data_source}{character: WT or PR8}

\item{sensitivity}{character: choice of "main" (model in main paper),
"n_L_n_I" (changing the number of latent and infectious stages), 
"loss" (ignoring loss of virus due to entry into target cells),
"mc" (use multi-cycle data only),
"linear_p" (production rate of virions increases linearly with time)}
}
\value{
a ggplot object. the bivariate density plot for R_0 and gen_time for multiple strains on the same plot
}
\description{
plot the bivariate density plot for R_0 and gen_time for multiple strains on the same plot
}
