% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_bivariate_density_single_strain}
\alias{plot_bivariate_density_single_strain}
\title{make Fig. S3}
\usage{
plot_bivariate_density_single_strain(strain)
}
\arguments{
\item{strain}{character: sH1N1, pH1N1, H5N1, H7N9 for the WT strains;
ENG195, Turkey05 for pH1N1-PR8 and H5N1-PR8 respectively}
}
\value{
a list of ggplot objects, each with one bivariate plot
}
\description{
make Fig. S3
}
