% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_cell_infect_param_histogram_by_strain}
\alias{plot_cell_infect_param_histogram_by_strain}
\title{Fig S3: plot histogram of cellular infection parameter}
\usage{
plot_cell_infect_param_histogram_by_strain(strain, sum_stat)
}
\arguments{
\item{strain}{character: sH1N1, pH1N1, H5N1, H7N9 for the WT strains;
ENG195, Turkey05 for pH1N1-PR8 and H5N1-PR8 respectively}

\item{cell_infect_param}{character: "gen_time", "log10_R_0" or "r".}
}
\value{
a ggplot object
}
\description{
Fig S3: plot histogram of cellular infection parameter
}
