% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_gen_dist_fixed_gen_time}
\alias{plot_gen_dist_fixed_gen_time}
\title{plot Fig. S5}
\usage{
plot_gen_dist_fixed_gen_time(constant_param)
}
\arguments{
\item{constant_param}{character vector: "log10_R_0", "gen_time" or "r"}
}
\value{
a ggplot object
}
\description{
plot Fig. S5
}
