% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_gen_dist_vary_params}
\alias{plot_gen_dist_vary_params}
\title{Make Fig. 4}
\usage{
plot_gen_dist_vary_params(constant_param)
}
\arguments{
\item{constant_param}{character: "gen_time" (A), "log10_R_0" (B), or "r" (C).
The parameter to hold constant.}
}
\value{
a ggplot object
}
\description{
Make Fig. 4
}
