% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_stats_across_strains}
\alias{plot_stats_across_strains}
\title{plot summary statistics or parameters across strains, for a particular set of
experiments}
\usage{
plot_stats_across_strains(sum_stat, sensitivity)
}
\arguments{
\item{sum_stat}{character vector of length 1.  if "cell_infect_params", plot generation
time, R_0 and initial growth rate.
if "mean_gen_no", plot the mean generation number.}

\item{sensitivity}{character: choice of "main" (model in main paper),
"n_L_n_I" (changing the number of latent and infectious stages), 
"loss" (ignoring loss of virus due to entry into target cells),
"mc" (use multi-cycle data only),
"linear_p" (production rate of virions increases linearly with time)}
}
\value{
ggplot object
}
\description{
plot summary statistics or parameters across strains, for a particular set of
experiments
}
