% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_viral_load}
\alias{plot_viral_load}
\title{make viral load plots in paper: Fig 1...}
\usage{
plot_viral_load(data_source, experiment, sensitivity)
}
\arguments{
\item{data_source}{character: WT or PR8}

\item{experiment}{character: "sc", "mc" or "mock" for single-cycle, 
multi-cycle and mock-yield experiments respectively}

\item{sensitivity}{character: choice of "main" (model in main paper),
"n_L_n_I" (changing the number of latent and infectious stages), 
"loss" (ignoring loss of virus due to entry into target cells),
"mc" (use multi-cycle data only),
"linear_p" (production rate of virions increases linearly with time)}
}
\value{
ggplot object
}
\description{
make viral load plots in paper: Fig 1...
}
