% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_for_inference.R
\name{read_and_process_all_data}
\alias{read_and_process_all_data}
\title{read and format data for single-cycle, multi-cycle and mock yield experiments}
\usage{
read_and_process_all_data(strain = "sH1N1-WT", inc_data)
}
\arguments{
\item{strain}{name of strain}

\item{inc_data}{character vector of up to length 5.  a subset of
\code{c("sc_inf", "sc_tot", "mc_inf", "mc_tot", "mock")} indicating which data
sets we're reading in}
}
\value{
list of data frames.  Each data frame is a data set, and has columns
"t" and "V_noise".  Each value of V_noise is a measurement corresponding to the time
in the same row.  If the value of V_noise is NA, that row is a placeholder
indicating that later on, we want to predict the viral load at that time.
}
\description{
read and format data for single-cycle, multi-cycle and mock yield experiments
}
