% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_for_inference.R
\name{read_PR8_data}
\alias{read_PR8_data}
\title{read PR8 data from prism file}
\usage{
read_PR8_data()
}
\value{
list of data frame with all data in file
}
\description{
read PR8 data from prism file
}
