% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_filenames.R
\name{set_dirs}
\alias{set_dirs}
\title{set directories to store computation results and plots}
\usage{
set_dirs(sensitivity)
}
\arguments{
\item{sensitivity}{character: choice of "main" (model in main paper),
"n_L_n_I" (changing the number of latent and infectious stages), 
"loss" (ignoring loss of virus due to entry into target cells),
"mc" (use multi-cycle data only),
"linear_p" (production rate of virions increases linearly with time)}
}
\value{
list with element "base": character: directory name for computation results
"figs": directory name for figures
}
\description{
set directories to store computation results and plots
}
