% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_filenames.R
\name{shorten_strain_name}
\alias{shorten_strain_name}
\title{remove data source from strain name}
\usage{
shorten_strain_name(strain)
}
\arguments{
\item{strain}{name of strain}
}
\value{
character vector of length 1
}
\description{
remove data source from strain name
}
