% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_summary_stats.R
\name{subset_summary_stats_from_rds}
\alias{subset_summary_stats_from_rds}
\title{from a summary stats .rds file, get a subset of the summary stats, undo log10 and thin}
\usage{
subset_summary_stats_from_rds(filename, my_summary_stats_names,
  n_samples = 1000, undo_log10 = TRUE)
}
\arguments{
\item{filename}{summary stats .rds file}

\item{my_summary_stats_names}{vector of summary stats to subset}

\item{n_samples}{number of samples to take.  If 0, take all samples}

\item{undo_log10}{if TRUE, for summary stats beggining with log10, take 10^
change summary stat name to reflect. if FALSE, do nothing}
}
\value{
subsetted and thinned matrix of summary stats
}
\description{
needed
}
