% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_for_inference.R
\name{transform_pars_wrapper}
\alias{transform_pars_wrapper}
\title{create a closure which transforms an unnamed numeric vector into
lists of named parameters to calculate the likelihood}
\usage{
transform_pars_wrapper(parTab, predict_data, G)
}
\arguments{
\item{predict_data}{character vector up to length 5.  subset of
\code{c("sc_inf", "sc_tot", "mc_inf", "mc_tot", "mock")}. indicates the
experiments about which we want to make predictions}

\item{G}{numeric vector: number of generations to track}
}
\value{
a function which transforms an unnamed numeric vector into
lists of named parameters to calculate the likelihood
}
\description{
create a closure which transforms an unnamed numeric vector into
lists of named parameters to calculate the likelihood
}
